Ahora se genera el archivo dirs, que contiene la lista de directorios ordenados por provincia segun aparecen en la tabla de los zip

+++ gen_urls.sh
@@ -1,6 +1,10 @@
 # Genera el listado de urls para descargar los ficheros automaticamente mediante wget
 cd temp
-rm urls
+
+if [ -f urls ]
+then
+        rm urls
+fi
 
 for num in `seq 2006 2008`
 do
